Refactor to add the SampleView plot that  extends the dictionary
